Split DB driver tests into generic & specific test case
We are going to replace the PgsqlDriver with a PDO
driver, so having tests beforehand that can be re-used
during the replacement will hopefully prevent issues
